apps/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
psycopg2-binary
